Implemented Glamourer logic to get designs, classes refactoring and creation of new folders
