+++ Noumenon/Gui/ConfigWindow.cs
@@ -1,4 +1,5 @@
 using System;
+using Dalamud.Configuration;
 using System.Numerics;
 using Dalamud.Interface.Windowing;
 using ImGuiNET;
@@ -7,7 +8,7 @@ namespace Noumenon.Windows;
 
 public class ConfigWindow : Window
 {
-    private Configuration configuration;
+    private Configuration.Configuration configuration;
 
     public ConfigWindow(Noumenon plugin) : base(
         "A Wonderful Configuration Window",
