+++ Noumenon/Noumenon.csproj
@@ -31,8 +31,7 @@
   <ItemGroup>
     <ProjectReference Include="..\ECommons\ECommons\ECommons.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Configuration\" />
-  </ItemGroup>
+    <ItemGroup>
+        <Folder Include="Configuration\" />
+    </ItemGroup>
 </Project>
